requirements.txt
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google.oauth
google-cloud-storage
google-cloud-secret-manager
python-dotenv